feat(QK-18): move husky and commitlint to root monorepo
