+++ apps/quanta-kit-vue-docs/eslint.config.mjs
@@ -1,37 +1,33 @@
+import { FlatCompat } from "@eslint/eslintrc";
+import js from "@eslint/js";
 import { dirname } from "path";
 import { fileURLToPath } from "url";
-import { FlatCompat } from "@eslint/eslintrc";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
 
 const compat = new FlatCompat({
   baseDirectory: __dirname,
+  recommendedConfig: js.configs.recommended,
 });
 
-const eslintConfig = [
+export default [
+  {
+    ignores: ["node_modules/**", "dist/**", "build/**"],
+  },
   ...compat.extends("plugin:vue/vue3-essential", "eslint:recommended"),
   {
+    files: ["**/*.js"],
     languageOptions: {
-      parser: "@babel/eslint-parser",
-      parserOptions: {
-        requireConfigFile: false,
-        babelOptions: {
-          babelrc: false,
-          configFile: false,
-          presets: ["@babel/preset-env"]
-        }
-      },
       globals: {
-        node: true
-      }
+        module: "readonly",
+        require: "readonly",
+        process: "readonly",
+        __dirname: "readonly",
+        __filename: "readonly",
+        exports: "writable",
+        global: "readonly",
+      },
     },
-    ignores: [
-      "node_modules/**",
-      "dist/**",
-      "build/**",
-    ],
   },
 ];
-
-export default eslintConfig;
