+++ apps/quanta-kit-vue-docs/package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "dev": "vue-cli-service serve",
     "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
+    "lint": "eslint . --ext .js,.vue,.ts --fix"
   },
   "dependencies": {
     "core-js": "^3.8.3",
@@ -15,12 +15,13 @@
     "@babel/core": "^7.12.16",
     "@babel/eslint-parser": "^7.12.16",
     "@babel/preset-env": "^7.12.16",
+    "@eslint/eslintrc": "^3",
+    "@eslint/js": "^9",
     "@vue/cli-plugin-babel": "~5.0.0",
     "@vue/cli-plugin-eslint": "~5.0.0",
     "@vue/cli-service": "~5.0.0",
     "eslint": "^9",
-    "eslint-plugin-vue": "^9.0.0",
-    "@eslint/eslintrc": "^3"
+    "eslint-plugin-vue": "^9.0.0"
   },
   "browserslist": [
     "> 1%",
