--- packages/quanta-kit-angular/.husky/pre-commit
@@ -1,8 +1,0 @@
-# Pre-commit checks
-echo "🔍 Running pre-commit checks..."
-
-# 1. Check TypeScript compilation
-echo "🔧 Checking TypeScript compilation..."
-npm run build
-
-echo "✅ All pre-commit checks passed!"
