+++ packages/quanta-kit-angular/package.json
@@ -14,8 +14,7 @@
     "storybook": "ng run quanta-kit:storybook",
     "build-storybook": "ng run quanta-kit:build-storybook",
     "test-storybook": "test-storybook --url http://localhost:6006 --config-dir projects/quanta-kit/.storybook --testTimeout 60000",
-    "semantic-release": "semantic-release",
-    "prepare": "husky"
+    "semantic-release": "semantic-release"
   },
   "prettier": {
     "overrides": [
@@ -47,8 +46,6 @@
     "@angular/cli": "^20.1.6",
     "@angular/compiler-cli": "^20.1.7",
     "@angular/platform-browser-dynamic": "^20.1.7",
-    "@commitlint/cli": "^19.8.1",
-    "@commitlint/config-conventional": "^19.8.1",
     "@compodoc/compodoc": "^1.1.26",
     "@semantic-release/changelog": "^6.0.3",
     "@semantic-release/commit-analyzer": "^13.0.0",
@@ -63,7 +60,6 @@
     "@storybook/builder-vite": "^9.1.3",
     "@storybook/test-runner": "^0.23.0",
     "@types/jest": "^29.5.14",
-    "husky": "^9.1.7",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
     "jest-preset-angular": "^14.6.1",
